+++ src/main/java/nl/nielspoldervaart/gdmc/GdmcHttpMod.java
@@ -1,10 +1,9 @@
-package com.gdmc.httpinterfacemod;
+package nl.nielspoldervaart.gdmc;
 
-import com.gdmc.httpinterfacemod.config.GdmcHttpConfig;
-import com.gdmc.httpinterfacemod.utils.RegistryHandler;
-import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.server.MinecraftServer;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.entity.player.PlayerEvent.PlayerLoggedInEvent;
 import net.minecraftforge.event.server.ServerStartingEvent;
 import net.minecraftforge.event.server.ServerStoppingEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -13,14 +12,18 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import nl.nielspoldervaart.gdmc.config.GdmcHttpConfig;
+import nl.nielspoldervaart.gdmc.utils.Feedback;
+import nl.nielspoldervaart.gdmc.utils.RegistryHandler;
 
 import java.io.IOException;
 
 // The value here should match an entry in the META-INF/mods.toml file
-@Mod("gdmchttp")
+@Mod("gdmchttpinterface")
 public class GdmcHttpMod
 {
-    public static final String MODID = "gdmchttp";
+    @SuppressWarnings("unused")
+    public static final String MODID = "gdmchttpinterface";
 
     // Directly reference a log4j logger.
     private static final Logger LOGGER = LogManager.getLogger();
@@ -42,10 +45,10 @@ public class GdmcHttpMod
 
         try {
             GdmcHttpServer.startServer(minecraftServer);
-            minecraftServer.sendSystemMessage(Component.nullToEmpty("GDMC Server started successfully."));
+            minecraftServer.sendSystemMessage(successMessage());
         } catch (IOException e) {
             LOGGER.warn("Unable to start server!");
-            minecraftServer.sendSystemMessage(Component.nullToEmpty("GDMC Server failed to start!"));
+            minecraftServer.sendSystemMessage(failureMessage());
         }
     }
 
@@ -54,5 +57,20 @@ public class GdmcHttpMod
         LOGGER.info("Server stopping");
 
         GdmcHttpServer.stopServer();
+    }
+
+    @SubscribeEvent
+    public void onPlayerLogIn(PlayerLoggedInEvent event) {
+        if (event.getEntity() != null) {
+            event.getEntity().sendSystemMessage(GdmcHttpServer.hasHtppServerStarted ? successMessage() : failureMessage());
+        }
+    }
+
+    private static MutableComponent successMessage() {
+        return Feedback.chatMessage("Server started at ").append(Feedback.copyOnClickText(String.format("http://localhost:%s/", GdmcHttpServer.getCurrentHttpPort())));
+    }
+
+    private static MutableComponent failureMessage() {
+        return Feedback.chatMessage("GDMC-HTTP server failed to start!");
     }
 }