+++ src/main/java/nl/nielspoldervaart/gdmc/GdmcHttpServer.java
@@ -1,9 +1,9 @@
-package com.gdmc.httpinterfacemod;
+package nl.nielspoldervaart.gdmc;
 
-import com.gdmc.httpinterfacemod.config.GdmcHttpConfig;
-import com.gdmc.httpinterfacemod.handlers.*;
 import com.sun.net.httpserver.HttpServer;
 import net.minecraft.server.MinecraftServer;
+import nl.nielspoldervaart.gdmc.config.GdmcHttpConfig;
+import nl.nielspoldervaart.gdmc.handlers.*;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
@@ -11,6 +11,8 @@ import java.net.InetSocketAddress;
 public class GdmcHttpServer {
     private static HttpServer httpServer;
     private static MinecraftServer mcServer;
+
+    public static boolean hasHtppServerStarted = false;
 
     public static int getHttpServerPortConfig() {
         return GdmcHttpConfig.HTTP_INTERFACE_PORT.get();
@@ -26,23 +28,25 @@ public class GdmcHttpServer {
         startServer(getHttpServerPortConfig());
     }
 
-    public static void startServer(int portNumber) throws IOException {
+    private static void startServer(int portNumber) throws IOException {
         // Create HTTP server on localhost with the port number defined in the config file.
         httpServer = HttpServer.create(new InetSocketAddress(portNumber), 0);
         httpServer.setExecutor(null); // creates a default executor
         createContexts();
         httpServer.start();
+        hasHtppServerStarted = true;
     }
 
     public static void stopServer() {
         if (httpServer != null) {
             httpServer.stop(5);
         }
+        hasHtppServerStarted = false;
     }
 
     private static void createContexts() {
         httpServer.createContext("/command", new CommandHandler(mcServer));
-        httpServer.createContext("/chunks", new ChunkHandler(mcServer));
+        httpServer.createContext("/chunks", new ChunksHandler(mcServer));
         httpServer.createContext("/blocks", new BlocksHandler(mcServer));
         httpServer.createContext("/buildarea", new BuildAreaHandler(mcServer));
         httpServer.createContext("/version", new MinecraftVersionHandler(mcServer));