+++ src/main/java/nl/nielspoldervaart/gdmc/commands/GetHttpInterfacePort.java
@@ -1,29 +1,26 @@
-package com.gdmc.httpinterfacemod.commands;
+package nl.nielspoldervaart.gdmc.commands;
 
-import com.gdmc.httpinterfacemod.GdmcHttpServer;
+import nl.nielspoldervaart.gdmc.GdmcHttpServer;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.context.CommandContext;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
-import net.minecraft.network.chat.Component;
+import nl.nielspoldervaart.gdmc.utils.Feedback;
 
 public class GetHttpInterfacePort {
 
 	private static final String COMMAND_NAME = "gethttpport";
-	private GetHttpInterfacePort() {}
 
 	public static void register(CommandDispatcher<CommandSourceStack> dispatcher) {
 		dispatcher.register(
 			Commands.literal(COMMAND_NAME)
-			.executes(context -> perform(context))
+			.executes(GetHttpInterfacePort :: perform)
 		);
 	}
 
 	private static int perform(CommandContext<CommandSourceStack> commandSourceContext) {
 		int currentPort = GdmcHttpServer.getCurrentHttpPort();
-		commandSourceContext.getSource().sendSuccess(Component.nullToEmpty(
-			String.valueOf(currentPort)
-		), true);
+		commandSourceContext.getSource().sendSuccess(() -> Feedback.chatMessage("Current GDMC-HTTP port: ").append(Feedback.copyOnClickText(String.valueOf(currentPort))), true);
 		return currentPort;
 	}
 }