+++ src/main/java/nl/nielspoldervaart/gdmc/utils/RegistryHandler.java
@@ -1,8 +1,8 @@
-package com.gdmc.httpinterfacemod.utils;
+package nl.nielspoldervaart.gdmc.utils;
 
-import com.gdmc.httpinterfacemod.commands.GetHttpInterfacePort;
-import com.gdmc.httpinterfacemod.commands.SetBuildAreaCommand;
-import com.gdmc.httpinterfacemod.commands.SetHttpInterfacePort;
+import nl.nielspoldervaart.gdmc.commands.GetHttpInterfacePort;
+import nl.nielspoldervaart.gdmc.commands.SetBuildAreaCommand;
+import nl.nielspoldervaart.gdmc.commands.SetHttpInterfacePort;
 import com.mojang.brigadier.CommandDispatcher;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraftforge.event.server.ServerStartingEvent;